requirements.txt
appdirs==1.4.4
attrs==21.2.0
bingads==13.0.10
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
cryptography==3.4.7
defusedxml==0.7.1
fiscalyear==0.3.2
future==0.18.2
google-ads==12.0.0
google-api-core==1.30.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
grpcio==1.38.1
httpretty==1.1.3
idna==3.2
isodate==0.6.0
lxml==4.6.3
mock==4.0.3
mysql-connector-python==8.0.25
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rich==10.4.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.19
suds-jurko==0.6
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
xmltodict==0.12.0
zeep==4.0.0
datetime-quarter
